refactor(console): split the e2e test case according to different scenarios
